// src/apis/models/ConsumerName.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Model for a generic consumer name
 * @export
 * @interface ConsumerName
 */
export interface ConsumerName {
  /**
   * First name of the user
   * @type {string}
   * @memberof ConsumerName
   */
  firstName: string;
  /**
   * Middle name of the user
   * @type {string}
   * @memberof ConsumerName
   */
  middleName?: string;
  /**
   * Last name of the user
   * @type {string}
   * @memberof ConsumerName
   */
  lastName?: string;
}

export function ConsumerNameFromJSON(json: any): ConsumerName {
  return ConsumerNameFromJSONTyped(json, false);
}

export function ConsumerNameFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConsumerName {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    firstName: json['firstName'],
    middleName: !exists(json, 'middleName') ? undefined : json['middleName'],
    lastName: !exists(json, 'lastName') ? undefined : json['lastName'],
  };
}

export function ConsumerNameToJSON(value?: ConsumerName | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    firstName: value.firstName,
    middleName: value.middleName,
    lastName: value.lastName,
  };
}