// src/apis/models/KYCStatus.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export type KYCStatus = 'SUCCESSFUL' | 'FAILED' | 'NOT_STARTED' | 'IN_PROGRESS' | 'MANUAL_REVIEW';